+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "helioslauncher",
-  "version": "1.7.0",
+  "version": "2.0.0-alpha.0",
   "productName": "Helios Launcher",
   "description": "Modded Minecraft Launcher",
   "author": "Daniel Scalzi (https://github.com/dscalzi/)",
