+++ src/main/config/model/LauncherConfig.ts
