+++ src/main/config/model/ModConfig.ts
