+++ src/main/config/model/NewsCache.ts
