+++ src/main/index.ts
@@ -4,7 +4,7 @@ import { join } from "path"
 import { readdirSync } from "fs-extra"
 import { format } from "url"
 import { autoUpdater } from 'electron-updater'
-import isdev from "./isdev"
+import isdev from "./util/isdev"
 
 const installExtensions = async () => {
     const installer = require('electron-devtools-installer');
