+++ src/main/mojang/model/auth/Agent.ts
