+++ src/main/mojang/model/auth/AuthPayload.ts
@@ -1,4 +1,4 @@
-import { Agent } from "./Agent";
+import { Agent } from './Agent'
 
 export interface AuthPayload {
