+++ src/main/mojang/model/auth/Session.ts
