+++ src/main/mojang/model/internal/Response.ts
