+++ src/main/mojang/model/internal/Status.ts
