+++ src/main/old/processbuilder.ts
@@ -4,14 +4,14 @@ import { join, basename } from 'path'
 import { ModuleWrapper, ServerWrapper } from './distromanager'
 import { Type, Required } from 'helios-distribution-types'
 import { LoggerUtil } from './loggerutil'
-import { ConfigManager } from './configmanager'
+import { ConfigManager } from '../config/configmanager'
 import { spawn } from 'child_process'
-import { SavedAccount } from './model/internal/config/SavedAccount'
+import { SavedAccount } from '../config/model/SavedAccount'
 import { tmpdir, release } from 'os'
-import { SubModConfig } from './model/internal/config/ModConfig'
+import { SubModConfig } from '../config/model/ModConfig'
 import { pseudoRandomBytes } from 'crypto'
 import { Util, LibraryInternal } from './assetguard'
-import { VersionJson, Rule } from './model/mojang/index/VersionJson'
+import { VersionJson, Rule } from '../asset/model/mojang/VersionJson'
 import { URL } from 'url'
 
 const logger = new LoggerUtil('%c[ProcessBuilder]', 'color: #003996; font-weight: bold')
